+++ layers/Engine/packages/component-model/src/FieldResolvers/AbstractQueryableFieldResolver.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace PoP\ComponentModel\FieldResolvers;
 
 use PoP\ComponentModel\Facades\ModuleProcessors\ModuleProcessorManagerFacade;
+use PoP\ComponentModel\FieldInterfaceResolvers\QueryableFieldInterfaceSchemaDefinitionResolverInterface;
 use PoP\ComponentModel\FieldResolvers\SelfQueryableFieldSchemaDefinitionResolverTrait;
 use PoP\ComponentModel\ModuleProcessors\FilterDataModuleProcessorInterface;
 use PoP\ComponentModel\Resolvers\QueryableFieldResolverTrait;
@@ -91,11 +92,14 @@ abstract class AbstractQueryableFieldResolver extends AbstractDBDataFieldResolve
     protected function convertFieldArgsToFilteringQueryArgs(TypeResolverInterface $typeResolver, string $fieldName, array $fieldArgs = []): array
     {
         $filteringQueryArgs = [];
-        if ($filterDataloadingModule = $this->getFieldDataFilteringModule($typeResolver, $fieldName)) {
-            $moduleprocessor_manager = ModuleProcessorManagerFacade::getInstance();
-            /** @var FilterDataModuleProcessorInterface */
-            $filterDataModuleProcessor = $moduleprocessor_manager->getProcessor($filterDataloadingModule);
-            $filterDataModuleProcessor->filterHeadmoduleDataloadQueryArgs($filterDataloadingModule, $filteringQueryArgs, $fieldArgs);
+        if ($schemaDefinitionResolver = $this->getSchemaDefinitionResolverForField($typeResolver, $fieldName)) {
+            /** @var QueryableFieldInterfaceSchemaDefinitionResolverInterface $schemaDefinitionResolver */
+            if ($filterDataloadingModule = $schemaDefinitionResolver->getFieldDataFilteringModule($fieldName)) {
+                $moduleprocessor_manager = ModuleProcessorManagerFacade::getInstance();
+                /** @var FilterDataModuleProcessorInterface */
+                $filterDataModuleProcessor = $moduleprocessor_manager->getProcessor($filterDataloadingModule);
+                $filterDataModuleProcessor->filterHeadmoduleDataloadQueryArgs($filterDataloadingModule, $filteringQueryArgs, $fieldArgs);
+            }
         }
         return $filteringQueryArgs;
     }