+++ layers/Schema/packages/comment-mutations-wp/src/TypeAPIs/CommentTypeMutationAPI.php
@@ -7,7 +7,7 @@ namespace PoPSchema\CommentMutationsWP\TypeAPIs;
 use PoP\ComponentModel\ErrorHandling\Error;
 use PoP\Translation\Facades\TranslationAPIFacade;
 use PoPSchema\CommentMutations\TypeAPIs\CommentTypeMutationAPIInterface;
-use PoPSchema\Comments\ComponentConfiguration as CommentsComponentConfiguration;
+use PoPSchema\CommentMutations\ComponentConfiguration;
 
 /**
  * Methods to interact with the Type, to be implemented by the underlying CMS
@@ -17,7 +17,7 @@ class CommentTypeMutationAPI implements CommentTypeMutationAPIInterface
     public function insertComment(array $comment_data): string | int | Error
     {
         // Convert the parameters
-        if (CommentsComponentConfiguration::mustUserBeLoggedInToAddComment()) {
+        if (ComponentConfiguration::mustUserBeLoggedInToAddComment()) {
             if (isset($comment_data['userID'])) {
                 $comment_data['user_id'] = $comment_data['userID'];
                 unset($comment_data['userID']);
@@ -31,13 +31,13 @@ class CommentTypeMutationAPI implements CommentTypeMutationAPIInterface
             $comment_data['comment_author_email'] = $comment_data['authorEmail'];
             unset($comment_data['authorEmail']);
         }
-        if (isset($comment_data['author-URL'])) {
-            $comment_data['comment_author_url'] = $comment_data['author-URL'];
-            unset($comment_data['author-URL']);
+        if (isset($comment_data['authorURL'])) {
+            $comment_data['comment_author_url'] = $comment_data['authorURL'];
+            unset($comment_data['authorURL']);
         }
-        if (isset($comment_data['author-IP'])) {
-            $comment_data['comment_author_IP'] = $comment_data['author-IP'];
-            unset($comment_data['author-IP']);
+        if (isset($comment_data['authorIP'])) {
+            $comment_data['comment_author_IP'] = $comment_data['authorIP'];
+            unset($comment_data['authorIP']);
         }
         if (isset($comment_data['agent'])) {
             $comment_data['comment_agent'] = $comment_data['agent'];